requirements.txt
pyyaml
py-trello
pytest
openpyxl
pandas
jira
python-dateutil
numpy
plotly
statsmodels
dash
setuptools
wheel
dash_bootstrap_components
dash-extensions
